presentorとcontainerに分けた実装
